+++ cyclelourcity-web-server/src/main/java/org/cycleourcity/cyclelourcity_web_server/datatype/SimplifiedTripEdge.java
@@ -2,17 +2,17 @@ package org.cycleourcity.cyclelourcity_web_server.datatype;
 
 public class SimplifiedTripEdge {
 	
-	private final int streetEdgeID;
+	private final double streetEdgeID;
 	private final String geometry;
 	private boolean bicycleMode;
 	
-	public SimplifiedTripEdge(int streetEdgeID, String geometry, boolean bicycle){
+	public SimplifiedTripEdge(double streetEdgeID, String geometry, boolean bicycle){
 		this.streetEdgeID = streetEdgeID;
 		this.geometry = geometry;
 		this.bicycleMode = bicycle;
 	}
 
-	public int getStreetEdgeID(){ return this.streetEdgeID; }
+	public double getStreetEdgeID(){ return this.streetEdgeID; }
 	
 	public String getGeometry(){ return this.geometry; }
